move installBinaries to main.ts, remove redundant runtimeDependencies.ts
